requirements/base.txt
fastapi==0.103.0
dependency_injector==4.41.0
requests==2.32.3
black==24.1.0
ruff==0.1.14
httpx==0.27.0
pandas==2.2.2
pydantic-settings==2.0.0